Fix broken scp (and possibly others) completions
The ssh alias to set the window title breaks scp completions
(and possibly others): they use the ssh command internally
(e.g. to retrieve the list of remote paths), and our
customisations breaks them.
Solve by checking the executed command, which should be 'ssh'
only in case of direct ssh command launched in the shell,
and bypass the alias if different command.